Big fixes and updates
the "database.js" file had part of its code reformed and fixed, and also had the "getArticles()" function added to return the complete list of articles.

Now, in the "app.js" file, error handlers log the error to facilitate solutions. "app.js" also had its functions changed so that instead of calling "database.articles" it calls "database.getArticles()" to avoid problems with getting an outdated list of articles and also for code aesthetics.